chore: add shared modal folder
