+++ src/components/Publication/Actions/Collect/CollectModule.tsx
@@ -3,10 +3,10 @@ import { gql, useMutation, useQuery } from '@apollo/client';
 import { PUBLICATION_REVENUE_QUERY } from '@components/Publication/Crowdfund';
 import { ALLOWANCE_SETTINGS_QUERY } from '@components/Settings/Allowance';
 import AllowanceButton from '@components/Settings/Allowance/Button';
-import Collectors from '@components/Shared/Collectors';
 import IndexStatus from '@components/Shared/IndexStatus';
 import Loader from '@components/Shared/Loader';
 import Markup from '@components/Shared/Markup';
+import Collectors from '@components/Shared/Modal/Collectors';
 import ReferenceAlert from '@components/Shared/ReferenceAlert';
 import ReferralAlert from '@components/Shared/ReferralAlert';
 import Uniswap from '@components/Shared/Uniswap';
