+++ src/components/Shared/Modal/Collectors.tsx
@@ -1,24 +1,25 @@
 import { gql, useQuery } from '@apollo/client';
 import UserProfile from '@components/Shared/UserProfile';
+import WalletProfile from '@components/Shared/WalletProfile';
 import { EmptyState } from '@components/UI/EmptyState';
 import { ErrorMessage } from '@components/UI/ErrorMessage';
 import { Spinner } from '@components/UI/Spinner';
-import { WhoReactedResult } from '@generated/types';
+import { Wallet } from '@generated/types';
 import { ProfileFields } from '@gql/ProfileFields';
-import { HeartIcon } from '@heroicons/react/outline';
+import { CollectionIcon } from '@heroicons/react/outline';
 import { Mixpanel } from '@lib/mixpanel';
 import { FC } from 'react';
 import { useInView } from 'react-cool-inview';
 import { PAGINATION } from 'src/tracking';
 
-import Loader from './Loader';
+import Loader from '../Loader';
 
-const LIKES_QUERY = gql`
-  query Likes($request: WhoReactedPublicationRequest!) {
-    whoReactedPublication(request: $request) {
+const COLLECTORS_QUERY = gql`
+  query Collectors($request: WhoCollectedPublicationRequest!) {
+    whoCollectedPublication(request: $request) {
       items {
-        reactionId
-        profile {
+        address
+        defaultProfile {
           ...ProfileFields
           isFollowedByMe
         }
@@ -36,13 +37,13 @@ interface Props {
   pubId: string;
 }
 
-const Likes: FC<Props> = ({ pubId }) => {
-  const { data, loading, error, fetchMore } = useQuery(LIKES_QUERY, {
+const Collectors: FC<Props> = ({ pubId }) => {
+  const { data, loading, error, fetchMore } = useQuery(COLLECTORS_QUERY, {
     variables: { request: { publicationId: pubId, limit: 10 } },
     skip: !pubId
   });
 
-  const pageInfo = data?.whoReactedPublication?.pageInfo;
+  const pageInfo = data?.whoCollectedPublication?.pageInfo;
   const { observe } = useInView({
     onEnter: () => {
       fetchMore({
@@ -54,20 +55,20 @@ const Likes: FC<Props> = ({ pubId }) => {
           }
         }
       });
-      Mixpanel.track(PAGINATION.LIKES);
+      Mixpanel.track(PAGINATION.COLLECTORS);
     }
   });
 
   if (loading) {
-    return <Loader message="Loading likes" />;
+    return <Loader message="Loading collectors" />;
   }
 
-  if (data?.whoReactedPublication?.items?.length === 0) {
+  if (data?.whoCollectedPublication?.items?.length === 0) {
     return (
       <div className="p-5">
         <EmptyState
-          message={<span>No likes.</span>}
-          icon={<HeartIcon className="w-8 h-8 text-brand" />}
+          message={<span>No collectors.</span>}
+          icon={<CollectionIcon className="w-8 h-8 text-brand" />}
           hideCard
         />
       </div>
@@ -76,21 +77,25 @@ const Likes: FC<Props> = ({ pubId }) => {
 
   return (
     <div className="overflow-y-auto max-h-[80vh]">
-      <ErrorMessage className="m-5" title="Failed to load likes" error={error} />
+      <ErrorMessage className="m-5" title="Failed to load collectors" error={error} />
       <div className="space-y-3">
         <div className="divide-y dark:divide-gray-700">
-          {data?.whoReactedPublication?.items?.map((like: WhoReactedResult) => (
-            <div className="p-5" key={like?.reactionId}>
-              <UserProfile
-                profile={like?.profile}
-                showBio
-                showFollow
-                isFollowing={like?.profile?.isFollowedByMe}
-              />
+          {data?.whoCollectedPublication?.items?.map((wallet: Wallet) => (
+            <div className="p-5" key={wallet?.address}>
+              {wallet?.defaultProfile ? (
+                <UserProfile
+                  profile={wallet?.defaultProfile}
+                  showBio
+                  showFollow
+                  isFollowing={wallet?.defaultProfile?.isFollowedByMe}
+                />
+              ) : (
+                <WalletProfile wallet={wallet} />
+              )}
             </div>
           ))}
         </div>
-        {pageInfo?.next && data?.whoReactedPublication?.items?.length !== pageInfo?.totalCount && (
+        {pageInfo?.next && data?.whoCollectedPublication?.items?.length !== pageInfo?.totalCount && (
           <span ref={observe} className="flex justify-center p-5">
             <Spinner size="md" />
           </span>
@@ -100,4 +105,4 @@ const Likes: FC<Props> = ({ pubId }) => {
   );
 };
 
-export default Likes;
+export default Collectors;
